add in-place versions of Von zur Gathen-Gerhard additive DFTs
These new functions use only one buffer for input and output which is extremly useful for large DFTs. They also use a temporary buffer of only 1/4 of the input buffer size, as opposed to 1 times the input size for the previous versions. For instance, the memory usage of a full 32-bit DFT is now 20GB (16GB data + 4GB buffer) as opposed to 48GB (3 x 16GB) before.
Speed is identical or better than previous versions.
previous routines will be removed in later commit.